Implemented radial scaled analog deadzone
The camera controls are much enhanced if x_cam and y_cam are floats,
because the joystick has to be pushed quite far just to reach an
integer value of 1, with the applied factor of 0.008.

In addition, I implemented a radial scaled deadzone which improves
analog joystick controls, because the motion will not 'get stuck' along
the axis for large tilts of the joysticks.

Finally, I changed analog joystick range to WIDE to allow for more
fine-grained controls, and reduced the analog deadzone for the camera
from 50 to 30